+++ src/main/java/ds/familydoctor/service/RequestService.java
@@ -2,8 +2,10 @@ package ds.familydoctor.service;
 
 import ds.familydoctor.entity.Request;
 import ds.familydoctor.repository.RequestRepository;
+import jakarta.persistence.EntityNotFoundException;
 import jakarta.transaction.Transactional;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.parameters.P;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -16,27 +18,31 @@ public class RequestService {
 
     @Transactional
     public Request getRequest(Long requestId) {
-        return reqRepo.findById(requestId).get();
+        return reqRepo.findById(requestId)
+                .orElseThrow(() -> new EntityNotFoundException("Request not found with this id: " + requestId));
     }
 
     @Transactional
-    public List<Request> getRequests() {
+    public List<Request> getAllRequests() {
         return reqRepo.findAll();
     }
 
     @Transactional
-    public void saveRequest(Request req) {
+    public Request save(Request req) {
         req.setStatus(Request.RequestStatus.PENDING);
-        reqRepo.save(req);
+        return reqRepo.save(req);
     }
 
     @Transactional
-    public void updateRequest(Request req) {
-        reqRepo.save(req);
+    public Request update(Request req) {
+        return reqRepo.save(req);
     }
 
     @Transactional
     public void deleteRequest(Long requestId) {
+        if (!reqRepo.existsById(requestId)) {
+            throw new EntityNotFoundException("Request not found with this id: " + requestId);
+        }
         reqRepo.deleteById(requestId);
     }
