+++ src/main/java/ds/familydoctor/service/UserService.java
@@ -2,6 +2,7 @@ package ds.familydoctor.service;
 
 import ds.familydoctor.entity.User;
 import ds.familydoctor.repository.UserRepository;
+import jakarta.persistence.EntityNotFoundException;
 import jakarta.transaction.Transactional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -16,26 +17,30 @@ public class UserService {
 
     @Transactional
     public User getUser(Long userId) {
-        return userRepo.findById(userId).get();
+        return userRepo.findById(userId)
+                .orElseThrow(() -> new EntityNotFoundException("User not found with this id: " + userId));
     }
 
     @Transactional
-    public List<User> getUsers() {
+    public List<User> getAllUsers() {
         return userRepo.findAll();
     }
 
     @Transactional
-    public void saveUser(User user) {
-        userRepo.save(user);
+    public User save(User user) {
+        return userRepo.save(user);
     }
 
     @Transactional
-    public void updateUser(User user) {
-        userRepo.save(user);
+    public User update(User user) {
+        return userRepo.save(user);
     }
 
     @Transactional
-    public void deleteUser(Long userId) {
+    public void delete(Long userId) {
+        if (!userRepo.existsById(userId)) {
+            throw new EntityNotFoundException("User not found with this id: " + userId);
+        }
         userRepo.deleteById(userId);
     }
